+++ lib/src/constants/mp_constants.dart
@@ -188,6 +188,8 @@ const double mp1OverPi = 1 / math.pi;
 const double mpDegreesInCircle = 360.0;
 
 const double mpLineSimplifyEpsilonOnScreen = 1.0;
+const int mpInteractiveLineSimplificationInitialIntensity = 1;
+const int mpInteractiveLineSimplificationMaxIntensity = 40;
 const double mpJoinLineExtremityToleranceOnScreen = 3.0;
 const double mpJoinLineExtremityToleranceOnScreenSquared =
     mpJoinLineExtremityToleranceOnScreen * mpJoinLineExtremityToleranceOnScreen;