+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mixins/mp_th2_file_edit_page_simplify_line_mixin.dart
@@ -17,7 +17,7 @@ mixin MPTH2FileEditPageSimplifyLineMixin on MPTH2FileEditState {
         if (isCtrlPressed || isMetaPressed) {
           cleanOriginalSimplifiedLines = false;
           if (isAltPressed && isShiftPressed) {
-            /// TODO open line simplification dialog box.
+            th2FileEditController.openInteractiveLineSimplificationDialog();
           } else {
             final MPLineSimplificationMethod newLineSimplificationMethod;